requirements.txt
tensorflow==2.9.0
tensorflow.keras==2.9.0
Pillow==10.0.0
numpy==1.24.4
scikit-learn==1.3.1
tifffile == 2023.7.10